fix(gateway): address MCP PR review feedback
Clear stale MCP discovery refresh state in the admin UI and render immediate refresh error summaries.

Harden gateway-managed MCP upstream credential resolution and long-lived Streamable HTTP timeout handling.

Split MCP gateway integration tests out of the main gateway test module.
